Update ignition token rotation flow
Before this PR:
The ignition-server/tokensecret-controller would delete the token Secret after a UUID ttl expired eventhough it was containing same the version/config.

After this PR:
The token Secret is the same for the whole NodePool lifetime as far as the version/config pair does not change.
The ignition-server/tokensecret-controller manages token rotation by periodically creating a new token for the existing Secret and persisting it in the cache while storing the old one in the secret in a different key.

This has some positive consequences:
- Rotation happens every time a token is beyond half of its ttl enabling an overlapping period of ttl/2 at max with the old token.
Overlapping fixes racing scenarios where a Machine is created with a userdata pointing to a token that gets expired right before the Machine have the chance to boot and fetch ign config.

- During token rotation there's no need to invoke the ignition-provider implementation to generate a new payload since we are only rotating the token but the value must remain the same for the lifecycle of the token Secret.