ensure no capi cluster is created for ibmcloud platformTypes Classic and VPC
Currently, CAPI clusters are getting deployed for Classic and VPC clusters despite not being operational for those providerTypes yet. This PR stops that so testing of the general system in these environments in IBMCloud can be performed and operations behave appropriately (as is delete operations are getting blocked because nothing is removing the CAPI cluster). It also provides unit tests for when someone does look to add a CAPI system to one of these environments (ensures they modify the unit tests and add more based on what changes they are introducing.)