Azure: Add node management
This makes the neccesary changes to the hostedcluster and nodepool
controller as well as to the controlplane operator and hosted cluster
config operator to allow creation of azure nodes to succeed. These nodes
join the cluster and become ready. What is not part of this and will be
done as a follow-up is:
* DNS setup
* All Clusteroperators become ready
* Setting up the in-tree cloudprovider

Ref https://issues.redhat.com/browse/HOSTEDCP-242