Azure: Setup public DNS
So far, the Azure create command doesn't have any DNS-related options
implemented. This adds --base-domain flag for create cluster and create
infra, passes on the base-domain to the cluster and uses it to find the
id of the zone that hosts the base domain, to then also pass that on.