Fix RESTMapper used by HosteAPICache in Control Plane Operator
Before this change, the HostedAPICache controller was creating a cache
with the RESTMapper from the management cluster. This caused an error
initializing the cache if a ClusterVersion resource did not exist in the
management cluster but did exist in the guest cluster.

This commit replaces the code to create the cache directly with code to
create a controller-runtime cluster and then gets the cache from it.
This results in the proper RESTMapper (from the guest cluster) being
used for the cache.