AWS: Support all properties of shared credentials files
We currently create an aws config with a SharedCredentials credentials
provider. This provider unfortunately only supports shared crdentials
that have a static credential and no other setting.

The only code in the AWS sdk that supports the full set of settings for
a shared credentials file is part of the session package, which means
the shared credentials file has to be set up as part of the session.

Additionally, the region must also be set in the session and not the
config, because otherwise that region will be used for all aws services,
even the ones that are global. This in turn breaks authentication
methods that need to contact STS, as that is not a global service.