Let NodePool controller wait if infraID is empty
The NodePool controller assumes infraID as contractual info (particularly uses it for naming generation). This https://github.com/openshift/hypershift/pull/1055 introduced a race so MachineDepoyment could be created without the infraID in the name.

This PR let the NodePool controller wait until th infraID is set in the HC API.