Factor out repeated code from delete-if-not-deleted code
This adds a new deleteIfNeeded that takes an arbitrary object, will try
to get it and only if that was successful and the object doesn't have a
deletionTimestamp yet it will be deleted. Any IsNotFound error along the
way is swallowed. All existing places where we had typed code for this
are updated to use the new func.