reduce frequency of leader election for components in the management cluster
This reduces the leader election frequencies of components that leader elect in the management cluster to the centralized API Server of that cluster. At scale: all these components end up aggregating requests to the API Server and if not reduced appropriately it can cause substantial load on both the etcd database and apiserver memory/cpu usage. The default values previously triggered election updates every 2 seconds. This pr moves those to be 15 seconds which should provide significant alleviation of the pressure the apiserver is experiencing at max scale (440 clusters in a management cluster).