Azure: Use random string rather than infraID as storage account name
The storage account is in the resource group so its name doesn't really
matter for cleanup. It has however some limitations wrt format, it must
be max 24 characters long and lower alphanumeric. This broke when we
added defaulting for the infraids, which also introduced a new format
that is too long and contains hyphens. This leads to failure:
https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_release/26666/rehearse-26666-periodic-ci-openshift-hypershift-main-periodics-e2e-conformance-azure/1499088629878755328#1:build-log.txt%3A79

What we do here matches what the installer does:
https://github.com/openshift/installer/blob/8fca1ade5b096d9b2cd312c4599881d099439288/data/data/azure/vnet/main.tf#L23