Add ServiceMonitor support for openshift-controller-manager
This commit adds ServiceMonitor support to the openshift-controller-manager
component, using the same label rules as OCP. It also refactors the existing
ServiceMonitor code so that they all share a single client certificate bound to
the `system:monitoring` role in the guest cluster. This was done to prevent
unnecessary secret and certificate sprawl.