add token expiration support to prevent nodepool upgrades from killing in flight provisions.
This pr enhances how ignition tokens are cleaned on nodepool upgrades. Previously on an upgrade: all ignition tokens were revoked as soon as the config was updated. The problem with this approach is outlined in issue: https://github.com/openshift/hypershift/issues/1020. To summarize: immediate revocation does not account for in flight provisions that were executed with the old token. This means these in flight operations would fail. This is a problem for consumers of the offering that should be addressed: provisions should never have the potential to fail. To fix this: when an upgrade occurs: old tokens are marked for expiration. The expiration time is based on the amount of time necessary to allow all in flight provisions to complete (new provisions will use the updated token). After that time: the ignition-server will proceed to delete the secret. The behavior on nodePool delete is unchanged: all tokens are revoked when the nodePool is deleted