Replace parameters in pool templates with the right RHCOS version
First the OpenStack RHCOS image, belonging to the current release is
detected. Then, if `{rhcos:openstack:checksum}` or
`{rhcos:openstack:url}` are found in the machine template, they are
replaced with the corresponding detected RHCOS version on machine
creation.

A new `ValidRHCOSImage` condition is added which has the same function
like `ValidAMI` but reports the  actual detected RHCOS version. It can
later be used by a broad variety of platforms, not just AWS.

Signed-off-by: Roman Mohr <rmohr@redhat.com>