fix dns reconciliation for ibm cloud or providers that provide a subdomain
Currently the DNS CRD instance is not getting set properly in IBM Cloud deployments. It is using the default value that I believe core Openshift expects where the cluster name is added after the base domain in the spec. However: that is not the case in IBM Cloud where the specified base domain is the base domain of the cluster (should not have the cluster preprended to it. We specify this through the DNS config already but hypershift is not respecting it and is always using the default value. In addition: Hypershift is currently missing the redirect standard OCP and ROKS deploys offer to the console from the master API endpoint. This enables that functionality by setting ConsolePublicURL appropriately.