Move oidc configmap management back into install command
When we moved this into the hypershift operator, the goal was to not
require credentials in the `install render` command. We don't get the
region and bucketname from credentials though, they are always passed in
so there is no reason why it can't be in the install command.

Having it there fixes some races where cluster creation fails because
the configmap does not exist yet because the operator didn't start yet.