Add ClusterID to HostedCluster API
Adds ClusterID field to HostedClusterSpec. The field value is generated
if left blank.
Modifies CVO so it can be initialized with the HostedCluster's ClusterID
before generating one.
The ClusterID is needed to add a replace label to ServiceMonitors and
PodMonitors so that metrics belonging to a particular cluster can be
identified.

Uses the clusterID in the HostedControlPlane to add a replace label to
every ServiceMonitor and PodMonitor created in a control plane namespace.