Allow overriding images at the hypershift operator level
Adds an image-refs flag to the install command that accepts an imagestream
YAML file that contains tags for images used by the HyperShift operator.
The images are then specified as environment variables on the hypershift
operator and propagated to the appropriate components.

Per-HostedCluster annotations still oveerride any operator-level environment
variables.