Add proxy support
This change adds proxy support. It adds a small controler into the
hypershift operator that makes it update its own deployement
based on the proxy resource in the cluster. It further extends the
hypershift operator to pass on its proxy env vars onto the capi provider
and the CPO and the CPO to pass them on to the KAS, KCM and HCCO and
ignition server.

Lastly, the konnektivity-socks5-proxy is extended to not route a
hardcoded list of cloudprovider domains into konnektivity but use its
own proxy configuration for that (if any).

Ref https://issues.redhat.com/browse/HOSTEDCP-331