dev: specify a more useful `ko` base image
Before this commit, images built with `ko` used a base image that included
no debugging/userland tools. This commit updates the base image to use the
"debug" variant of the base image which is nominally larger but includes
a busybox userland, which is far more useful by default for a low cost.