Introduce `OIDCConfigurationInvalid` condition for OIDC setup validation
This commit introduces a new `OIDCConfigurationInvalid` condition for
HostedCluster. If OIDC documents cannot be reconciled for a cluster,
`OIDCConfigurationInvalid=True`, otherwise `OIDCConfigurationInvalid` is
removed. The message associated with the condition tries to provide some more
helpful context about what might be wrong to help the user correct the issue.
More error interpretation can be added in the future.