Hypershift-operator: Increase worker count
This increase the workers for the hostedcluster operator to 10 (from 1)
and for the nodepool controller to 30 (also from 1). Otherwise the
workers become a bottleneck if more than one hostedcluster is changed at
a time.

Both the 10 workers and the 30 workers are relatively arbitrary chosen,
these numbers might need revisiting in the future or even warrant a
config option.