Registry configuration: reconcile only what we need to changes
In its current state, the hosted cluster config operator overwrites any
changes made by the guest cluster admin to the registry configuration.
This prevents changes like enabling a route or increasing the number of
replicas.

This commit limits what we change to things we need to change and leave
everything else as is.