Dump: Always create an archive
Currently, dump just drops a lot of files. This is useful for browsing
them in the CI job output, but terrible for downloading them for local
inspection, as downloading a lot of files is extremely slow, even if the
files aren't big.

This change makes us always create an archive of the dump to not require
extending every CI job to do this manually.