Create valid route names with long namespace names
When using a long name for a cluster namespace, the routes that the
control plane operator creates is invalid because a segment of the
domain name has a length > 63 chars.

This commit sets the host name for routes to ensure that the resulting
host name is valid. Modification of the route hostname only happens when
the resulting hostname will be invalid. The name is shortened by
trimming the original name and appending a hash of the entire name,
resulting in repeatable and distinct names every time the name is
shortened.