e2e: Don't fail test on transient recoverable API lookup
Before this commit, calls to `WaitForConditionsOnHostedControlPlane()` could
fail a test if an API lookup fails even though that lookup is recoverable and
retried automatically. This made the test flaky.

This commit fixes the code so that these retriable errors are logged but do
not fail the test.

This commit also moves a log message which was intended to emit during retries
but was instead placed at the exit point.