e2e: Don't enable user workload monitoring on management clusters
Before this commit, UWM was enabled by the e2e `setup` command, which
was used in the past but is no longer used. The UWM stack is thus wasting
resources on management clusters used for e2e runs.

This commit removes UWM from the monitoring setup for e2e tests.