Enforce hostedcluster service route immutability constraints
Before this PR, the `route.spec.host` field associated with a given
`hostedcluster.spec.services[].route` value was being defaulted by OCP and then
later overwritten with an empty string on subsequent reconcile loops. This is
fundamentally a side effect of a violation of the immutability constraint on the
`hostname` field.

To fix the issue, this PR refactors the relevant Route reconciliation code to
compute and set the `route.spec.host` field only during the creation of the
Route resource, which enforces the immutability constraint and resolves the side
effect.