Make HyperShift operator compatible with previous CPOs without utilities
Adds symlinks in HyperShift operator image for previously standalone
utilities and enables the control plane operator to invoke the right
command based on the symlink it was invoked with.

Checks whether the CPO for a given cluster includes subcommands for
utilities by checking the labels on the CPO image, and only uses the
CPO image for utilities if it does include the labels.