Fix reconcile loop in nodepool controller
Before this commit, the nodepool controller reconciled in a hotloop
while trying to apply a `nil` value to the machinedeployment annotations
field, a patch which is having no effect when executed through the CAPI
`PatchHelper` utility.

This commit fixes the issue in a temporary way by moving the offending
map assignment to a place where the map is only initialized when we also
set map keys. A better future fix would be to use the `CreateOrUpdateProvider`
which can deal with defaulting so that we don't have to solve the bug in
upstream utility code.