konnectivity ds: Tolerate all taints
The konnektivity DS already runs with hostnetwork in order to be up even
if CNI doesn't come up. It currently ends up not being scheduled in that
situation though, because it doesn't tolerate the not-ready taint.

Make it tolerate all tains, as this is very helpful especially when CNI
doesn't come up.