Add Metrics Set for Telemetry metrics
Introduces the concept of a metrics set with 3 initial sets defined:
- Telemetry: only produce metrics needed for telemetry
- SRE: produce metrics needed by an
  SRE team in charge of keeping the control planes functioning.
- All: produce all available metrics.

Using a much reduced metrics set (RemoteHealthMonitoring) makes it
possible to setup UWM to collect control plane metrics and forward them
to telemetry without incurring a large burden in terms of resources for
the UWM stack. Customers can update the setting to a larger set of
metrics if they are willing to dedicate more resources to metrics
collection.