Add test for NodePools in place upgrades
Enables e2e test for in place upgrades https://github.com/openshift/hypershift/pull/1258 so any follow up will be test driven.
As part of getting the test green this PR moves from querying the ign server to store the payload in the secret token so it can be consumed by the in place upgrade process. Otherwiswe the ign server is not reachable form the NodePool controller in private topologies.