Only apply hostedcluster updates when the resource has changed
Before this commit, hostedcluster controller would perform client update
calls whether or not there was any change to the resource, resulting in
excessive no-op calls to the apiserver.

This commit adds semantic equality checks around the updates so that the
API calls are only made when the resource has changes to apply.