Revise and simplify NodePool's KubeVirt platform API
Previously, the KubeVirt platform's NodePool api consisted of the entire KubeVirt VM spec. We started with that because it gave us the freedom to quickly adapt our configurations as we experimented with various network and storage tunings. This does not provide a reasonable user experience though.

Now that we've gained more experience with the project, we feel comfortable pairing down the KubeVirt platform API to a small handful of tunings that are more approachable to end users.

The goal of this api revision is to provide a simplified api for configuring KubeVirt VMs within a Hypershift NodePool that only exposes the bare minimal tuning users need to be aware of when using the KubeVirt platform. We've also attempted to structure this API in a way that allows us to expand easily in the future, which is why there is some nested structure patterns that might not seem optimal at first glance.

For now, the API only has two top level structs `compute` and `rootVolume`. The compute structure lets users define things like the number of cores and amount of memory to give the guest. The `rootVolume` structure is where users define where the rhcos images comes from and the type of storage to use (only persistent is supported right now, but this can be expanded to ephemeral storage in the future)

In the future we might have a new top level struct dedicated to network configurations once more advanced tunings are supported there. Also, the compute struct might get the addtion of a VM `flavor` to define the mem/cpu usage for the guest once Openshift Virtualization officially supports flavors.

More info related to the api structure can be found in [this document](https://docs.google.com/document/d/17vMxzPZuq-XvHkkTp6E-CPdopMdNqOk48_0e61G_Pmg/edit#heading=h.y4i6eo4m47u3).

With this api change, we've transitioned the kubevirt platform to only use persistent storage (VM's backed by PVCs). As a result, we've introduced two new hypershift CLI options, `--root-volume-size` and `--root-volume-storage-class`. Neither values have are required, but offer users the ability to customize the root volume size and the storage class that will provision the volumes.

None of these new options are required for the hypershift cli though. Using arguments that previously worked with hypershift cli will continue to work even after this change.

```
spec:
    clusterName: example-6thkl
    management:
      autoRepair: false
      replace:
        rollingUpdate:
          maxSurge: 1
          maxUnavailable: 0
        strategy: RollingUpdate
      upgradeType: Replace
    nodeCount: 5
    platform:
      kubevirt:
        compute:
          cores: 2
          memory: 4Gi
        rootVolume:
          diskImage:
            containerDiskImage: quay.io/dvossel/rhcos:4.11-pre-release
          persistent:
            size: 16Gi
      type: KubeVirt
    release:
      image: quay.io/openshift-release-dev/ocp-release:4.11.0-0.nightly-multi-2022-04-19-151226-x86_64
```

Signed-off-by: David Vossel <davidvossel@gmail.com>