Omit MachineConfig and MachineConfigPool CRDs from release payload
The MachineConfig and MachineConfigPool CRDs are not needed in a
HyperShift cluster and they should be removed from the UI.

This change is to omit them from the release payload applied to the
guest cluster. It also disables the node tuning operator because it uses
MachineConfig resources to work.

The node tuning operator can later be enabled if it's modified to work
with HyperShift clusters.