Refactor HostedCluster AWS role API
Before this commit, the HostedCluster AWS role API was considered difficult to
understand while also leaking implementation details to the end user. The user
is actually responsible for creating specific required AWS policy documents and
providing the ARNs for those roles as API input. Before this change, the user
was required to supply not only the ARNs, but also the names and namespaces of
secrets the inputs are mapped to internally by HyperShift, which are
implementation details. This is a problem because not only are those constant
details irrelevant to the end user, they could change in a future HyperShift
release and thus introduce compatibility or migration problems.

This commit improves the API and implements HOSTEDCP-153 by refactoring the
opaque role array into a struct which requires only the ARNs for concrete roles
whose policy document requirements are documented on concrete fields. The
internal details of how these ARNs are used is no longer exposed to the end
user.