disable reconcile of registry config in IBMCloud deployments
Currently there is a race condition in IBMCloud deployments where if hypershift initializes the registry config before the managed service operation has a chance to properly initialize the registry config: It will result in a bad initial config that will cause the cluster-image-registry operator to crash and lead to tickets. For IBMCloud deployments: the initialization of this config should be delegated to the managed service to do the initalization so this does not occur. From there: the user can edit however they like in line with upstream openshift