Export `ign_server_payload_generation_seconds` metric and add e2e budget
Before this commit, the `ign_server_payload_generation_seconds` metric was not
being accounted for in the e2e metrics budget assertions, which led to us
failing to detect a possible performance regression on AWS with the new ignition
payload generation implementation.

This commit adds a recording rule for `ign_server_payload_generation_seconds`
(to ensure the metric is exported), and also implements a budget check in the
e2e fixture. For now, the budget is a p90 expectation of 45 seconds for payload
generation, which is very probably conservative. We've observed an intermittent
8 minute value in one known e2e test case that hasn't yet been reproduced. We
should adjust the histogram bucket and budget once we have more data around the
implementation's real world performance.

Hopefully this budget will help us detect problems and narrow down a reproducer
for the anomaly.