Change webhook to implement webhook.CustomValidator
This give us the ability to:
scope different webhooks to packages/files so the main does not grow indefinitely.
Reusing controller runtime interface.
Drop unnecessary existing logic in main.
Give a consistent way to share validation logic from controllers which does not necessarily need to be exported.
i.e Sustainable code structure.

Since we want to implement the CR outside the the API package we create a mock Webhook struct for HostedCluster.