Move autoscaler/mapprover reconciliation down to CPO for >=4.11
Machine approver and autoscaler images are already inferred from the payload for two reasons:
- To prevent unnecessary rollouts from happening during HO upgrades.
- Get them closer with the OCP/Kube releases they have been tested with.

This PR move the reconciliation from the HC down to the CPO so the instantiation of the components is consistent with their version. This reduces the risk to introduce breaking changes in the HO for older releases images.
For < 4.11 we let the HO to keep reconciling to stay consistent with existing IBM clusters and satisfy validations. This is signaled via CPO Docker image labels.