Make CPO image configurable in `ci-test-e2e` Make target
This commit makes the CPO image configurable within the `ci-test-e2e`
make target via the `CONTROL_PLANE_OPERATOR_IMAGE` environment variable.

If the value is an empty string, the CPO image won't be overridden, which
makes this change compatible with the existing usage in the periodics while
providing the override for presubmits.