e2e: Increase parallel test limit
Before this commit, e2e test parallelism was limited to the default of
`GOMAXPROCS`, which was causing some of our tests to pause execution
so long that the tests can time out, even though those tests could have
executed in parallel.

This commit increases the parallelism limit so that all our tests run
in parallel.