HAProxy ignition config: Use nodepool release image
This makes us use the nodepool release image as source for the haproxy
image. To do this, the reconciliation of the haproxy ignition config is
moved from the CPO into the nodepool controller. Since the config is
then generated at the same place it is consumed, we do not write it to a
configmap anymore but just do the generation in memory.

Fixes https://issues.redhat.com/browse/HOSTEDCP-426