Change ignition server http code if token not found.
This fixes a racing issue between the payload being available in the ign-server and any machine requesting it and being able to boot.
Let's follow the path of *-7b2b7-us-east-1b-e731d66e which is the token secret referenced by a MachinSet which Machine was unable to boot.

The payload is generated at 2022-05-18T12:39:39Z","msg":"got ignition payload
https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/pr-logs/pull/openshift_hypershift/1399/pull-ci-openshift-hypershift-main-e2e-aws-all/1526895205024796672/artifacts/e2e-aws-all/test-e2e/artifacts/TestReplaceUpgradeNodePool_PreTeardownClusterDump/namespaces/e2e-clusters-9v26f-example-7b2b7/core/pods/logs/ignition-server-74bd6bd554-qr7z6-ignition-server.log

Right after multiple machines send http request to the ign-server
2022/05/18 12:39:33 User Agent: Ignition/2.13.0. Requested: /ignition
2022/05/18 12:39:33 User Agent: Ignition/2.13.0. Requested: /ignition
2022/05/18 12:39:34 User Agent: Ignition/2.13.0. Requested: /ignition

Including the one failing to boot
https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/pr-logs/pull/openshift_hypershift/1399/pull-ci-openshift-hypershift-main-e2e-aws-all/1526895205024796672/artifacts/e2e-aws-all/test-e2e/artifacts/TestReplaceUpgradeNodePool_PreTeardownClusterDump/machine-console-logs/example-7b2b7-us-east-1b-fk9g6.log

Because the ign-server is return a 4xx if the token required does not exist Ignition considers the process unsuccessful but complete without retrying and backoff https://coreos.github.io/ignition/operator-notes/#http-backoff-and-retry

This also sets a test function time out to avoid hitting the core CI platform global timeout, thi let us fail earlier  and dump the artifacts for the test.