Replace router shard with private router in control plane namespace
Before this commit, we were creating IngressController resources for
ingress from private clusters. This has the problem of preventing the
management cluster from upgrading because the ingress operator cannot
create corresponding DNS records for the IngressController.

This commit removes the code to create an IngressController and instead
creates a router pod (and load balancer service) inside the control
plane namespace, avoiding the issue with the ingress operator.