Add allowed principals to service endpoints
Before this commit creating a private link between a management cluster
and a hosted cluster in a different AWS account failed because the
service endpoints created on the management side did not allow endpoints
on the guest side to connect.

This PR adds the control plane role as an allowed principal to the
endpoint service, making it possible to establish the private link
across accounts.

The way it obtains the control plane role is by reading the control
plane credentials secret referenced by the hosted cluster. In order
to parse the secret (in ini format), it vendors the
gopkg.in/ini.v1 package.