adopt existing immutable selectors to prevent errors reconciling components from roks toolkit clusters
This fixes an error in the control plane operator where it will fail to reconcile and properly adopt deployments based on roks toolkit clusters. This is due to the fact that the selectors has changed and that is an immutable field. Specifically, hypershift selectors add a duplicate selector hypershit.component that is not contained in roks toolkit clusters. Even with the single app selector: deployment rollout behavior is unchanged. This allows for zero downtime adoption of roks toolkit clusters. Associated issue: https://github.com/openshift/hypershift/issues/1042