Allow nodepool deletion when hcluster is gone
Currently, nodepool deletion will fail if the hostedcluster doesn't
exist anymore which means that if nodepool deletion is too slow, we will
deadlock. This fixes that.

Note that this introduces an backwards-incompatible change to how
machinedeployments are named. This means that the hypershift operator
will create a new machinedeployment for existing nodepools when
upgraded. The same goes for cleanup, the machinedeployment will be
deleted transiently by the garbage collector controller, however the
same goes for the capi pods, hence cleanup might not work automatically.