Use git commit as version in HyperShift binaries
Adds git commit, tree state and build date as ldflags to hypershift
binaries. It then includes the version in output logs.
It also adds a command `version` that outputs the version from the CLI
and also includes that version as an annotation in the HyperShift
deployment to indicate which version of the CLI was used to generate it.

This enables us to better debug errors and understand which version of
the code is running in a particular installation.