Increase MHC nodeStartupTimeout
This is to try to mitigate and isolate CI flakiness https://github.com/openshift/hypershift/pull/1430#issuecomment-1144666152 where ocassionaly cluster creation with autorepair enabled has failed to get healthy Nodes. In these failures Machines are being deleted when the NodeStartupTimeout is met. If there's a legit reason why the Nodes are not being created this won't help, however it will help to make more clear the former is true.