Reduce HyperShift operator metrics cardinality
To include an initial set of metrics for telemetry, this PR reduces the
number of labels to report for hosted clusters and nodepools. This is
to keep an initial set of metrics under 10 series. (5 for hc, 5 for np)

The current metric with multiple labels is also flawed in that it can
potentially count a cluster multiple times as its version or endpoint
access type changes. We will need a way to update the counts for all
possible values of each label, but this can be done at a later time.