Drop kas call in agent ReconcileCAPIInfraCR and fail if no ign
This add unit tests and fixes the ReconcileCAPIInfraCR agent implementation to avoid an unncesesary kas call to fetch the HCP, since all needed info is available already.
Aditionally it returns early if there's no ign endpoint to ensure we create the agentCluster only after ignition endpoint exists so AgentClusterInstall is only created with the right ign to boot machines.
https://bugzilla.redhat.com/show_bug.cgi?id=2097895
https://github.com/openshift/assisted-service/blob/241ad46db74add5f16e153f5f7ba0a5496fb06ba/pkg/validating-webhooks/hiveextension/v1beta1/agentclusterinstall_admission_hook.go#L185
https://github.com/openshift/cluster-api-provider-agent/blob/master/controllers/agentcluster_controller.go#L126