install: add operator-only option for OCP metrics
Instead of providing a simple boolean for enabling OCP monitoring for
the HyperShift operator and all control plane namespaces, this commit
introduces an option to enable OCP monitoring only for the HyperShift
operator and not control plane namespaces.

This additional option is needed for the MCE standalone use case in
which we want to use OCP monitoring to collect metrics from the
HyperShift operator to send to telemetry. For control plane namespaces
we want to use User Workload Monitoring to remote write telemetry of the
individual hosted clusters.