Disable the Node Tuning Operator
As mentioned in commit b67f8bcd6262e66ae3eb92e21efb24f7aec6e1c4,
NTO currently depends on the MachineConfig / MachineConfigPool CRDs
and is therefore not compatible with HyperShift. Until NTO
is modified to run in HyperShift clusters, it should be disabled.
In this previous commit only the *-operator-ibm-cloud-managed.yaml
manifest was disabled, not the default operator deployment manifest.