Get rid of hostedapicache by using HCCO
Currently, the CPO needs the hostedapicache to access the clusterversion
of the guest cluster in order to manage two conditions on the hosted
controlplane. This results in always yielding some errors during startup
when the guest cluster is not yet available and is relatively
complicated code. Move the functionality into a new controller in the
HCCO and get rid of the hostedapi cache.