Remove .status from serialized configuration on HCP
In the typed configuration we store, we only use the spec struct of each
configuration resource. However, when we serialize it, we include the
entire resource. The stauts resource is not needed and is causing
incompatibility with older OCP versions.

Going forward, the incompatibility should be fixed by #1616. However,
this change is needed to ensure we don't immediately break things while
CPOs in production are upgraded.